+++ R/ambiorix.R
@@ -35,19 +35,19 @@
 #' @export
 scaffold_ambiorix <- function(vue = FALSE, use_cdn = TRUE, edit = interactive()){
   # checks
-  assert_that(has_npm())
+  assert_that(has_engine())
   assert_that(is_ambiorix())
 
   open_msg("ambiorix")
 
   # init npm
-  npm_init()
+  engine_init()
 
   # install dependencies
   webpack_install()
 
   # edit package.json
-  npm_add_scripts()
+  engine_add_scripts()
 
   # set up dir for golem
   # only create dir if vue or react