+++ R/extension.R
@@ -29,7 +29,7 @@
 #' @export
 scaffold_extension <- function(name, edit = interactive()){
   # checks
-  assert_that(has_npm())
+  assert_that(has_engine())
   assert_that(is_package())
   assert_that(not_missing(name))
   assert_that(is_name_valid(name))
@@ -40,7 +40,7 @@ scaffold_extension <- function(name, edit = interactive()){
   open_msg("shiny extension", name)
 
   # init npm
-  npm_init()
+  engine_init()
 
   # create base npm webpack files
   create_directory("srcjs/exts")
@@ -52,7 +52,7 @@ scaffold_extension <- function(name, edit = interactive()){
   webpack_install()
 
   # edit package.json
-  npm_add_scripts()
+  engine_add_scripts()
 
   # create config file
   configure(