+++ NEWS.md
@@ -8,6 +8,7 @@ callable from outside the bundle.
 - Add `modules` and `import` arguments to CSS loader function.
 - Add `use_loader_style` to easily allow CSS modules. See the
 [documentation](https://packer.john-coene.com/#/guide/style)
+- Add support for yarn via `engine_*` functions.
 
 ## packer 0.0.6